refactor: remove repeated type
