+++ app/scripts/lib/rpc-method-middleware/handlers/types.ts
@@ -1,4 +1,9 @@
 import { JsonRpcParams, JsonRpcRequest } from '@metamask/utils';
+import {
+  CaveatSpecificationConstraint,
+  PermissionController,
+  PermissionSpecificationConstraint,
+} from '@metamask/permission-controller';
 import { MessageType } from '../../../../../shared/constants/app';
 
 export type HandlerWrapper = {
@@ -10,3 +15,12 @@ export type HandlerRequestType<Params extends JsonRpcParams = JsonRpcParams> =
   Required<JsonRpcRequest<Params>> & {
     origin: string;
   };
+
+type AbstractPermissionController = PermissionController<
+  PermissionSpecificationConstraint,
+  CaveatSpecificationConstraint
+>;
+
+export type GrantedPermissions = Awaited<
+  ReturnType<AbstractPermissionController['requestPermissions']>
+>[0];
